Handling-exceptions

+++ src/main/java/com/example/reservation/service/DoctorService.java
@@ -1,14 +1,19 @@
 package com.example.reservation.service;
 
+import com.example.reservation.dto.AppointmentFromDoctorPovDTO;
 import com.example.reservation.dto.DoctorDTO;
+import com.example.reservation.exception_handler.CannotDeleteException;
+import com.example.reservation.model.Appointment;
 import com.example.reservation.model.Doctor;
 import com.example.reservation.repository.DoctorRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -18,6 +23,9 @@ public class DoctorService {
     private final DoctorRepository repository;
 
     public Optional<DoctorDTO> getDoctor(Integer id) {
+        if (!repository.existsById(id)) {
+            throw new IllegalArgumentException("Doctor not found");
+        }
         return repository.findById(id)
                 .map(DoctorDTO::new);
     }
@@ -38,7 +46,19 @@ public class DoctorService {
         return repository.save(doctor);
     }
 
-    public Optional<Doctor> deleteDoctor(int id) {
+    public Optional<Doctor> deleteDoctor(int id) throws CannotDeleteException {
+        if (!isDoctorExist(id)) {
+            throw new IllegalArgumentException("Doctor not found");
+        }
+        DoctorDTO doctor = getDoctor(id).orElse(null);
+        Set<AppointmentFromDoctorPovDTO> appointments = Optional.ofNullable(doctor.getAppointments())
+                .orElseGet(Collections::emptySet);
+        if (!appointments.isEmpty()) {
+            List<Integer> ids = appointments.stream()
+                    .map(AppointmentFromDoctorPovDTO::getAppointmentId)
+                    .collect(Collectors.toList());
+            throw new CannotDeleteException("Cannot delete Doctor due to appointment scheduled. Appointments id: " + ids);
+        }
         return repository.deleteById(id);
     }
